DFG's flushForTerminal() needs to add PhantomLocals for bytecode live locals. https://bugs.webkit.org/show_bug.cgi?id=160755 <rdar://problem/27488507>
Reviewed by Filip Pizlo.

JSTests:

* stress/need-bytecode-liveness-for-unreachable-blocks-at-dfg-time.js: Added.

Source/JavaScriptCore:

If the DFG sees that an inlined function will result in an OSR exit every time,
it will treat all downstream blocks as dead.  However, it still needs to keep
locals that are alive in the bytecode alive for the compiled function so that
those locals are properly written to the stack by the OSR exit ramp.

The existing code neglected to do this.  This patch remedies this issue.

* dfg/DFGByteCodeParser.cpp:
(JSC::DFG::ByteCodeParser::flushDirect):
(JSC::DFG::ByteCodeParser::addFlushOrPhantomLocal):
(JSC::DFG::ByteCodeParser::phantomLocalDirect):
(JSC::DFG::ByteCodeParser::flushForTerminal):



Identifier: 178871@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204360 268f45cc-cd09-0410-ab3c-d52691b4dbfc