Optimization in Node.insertBefore() is not spec-compliant https://bugs.webkit.org/show_bug.cgi?id=160746
Reviewed by Ryosuke Niwa.

LayoutTests/imported/w3c:

Rebaseline W3C test now that more checks are passing.

* web-platform-tests/dom/ranges/Range-mutations-expected.txt:

Source/WebCore:

We have an optimization in Node.insertBefore(newNode, refChild) to avoid
doing any work when newNode == refChild or newNode.nextSibling == refChild.

This optimization is not in the specification:
- https://dom.spec.whatwg.org/#concept-node-replace

The issue is that this optimization is observable with DOM mutation
observers / listeners or DOM ranges.

This patch addresses the issue by dropping the optimization. This case
does not seem common enough to be worth optimizing for. However, if it
turns out to regress the performance of things we care about, we could
fallback to doing the optimization only when it is not observable.

Test: fast/dom/Node/insertBefore-no-op-mutationobserver.html

* dom/ContainerNode.cpp:
(WebCore::checkAcceptChild):
Move refChild->parent() == parent check from insertBefore() to our
pre-insertion check function, right after checking if newNode contains
parent. This was done in order to match more closely the specification
and to make sure that exception are thrown in the correct order:
- https://dom.spec.whatwg.org/#concept-node-ensure-pre-insertion-validity (steps 2 and 3)

(WebCore::ContainerNode::insertBefore):
1. Drop the (refChild->previousSibling() == &newChild || refChild == &newChild)
   optimization as it is not spc-compliant.
2. If refChild is newNode, then set refChild to newChild.nextSibling as per:
   - https://dom.spec.whatwg.org/#concept-node-pre-insert (step 3)

LayoutTests:

Add layout test to make sure mutation observers / listeners are always
notified when Node.insertBefore() is called.

* fast/dom/Node/insertBefore-no-op-mutationobserver-expected.txt: Added.
* fast/dom/Node/insertBefore-no-op-mutationobserver.html: Added.


Identifier: 178879@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204368 268f45cc-cd09-0410-ab3c-d52691b4dbfc