The jsc shell's Element host constructor should throw if it fails to construct an object. https://bugs.webkit.org/show_bug.cgi?id=160773 <rdar://problem/27328608>
Reviewed by Saam Barati.

JSTests:

* stress/generational-opaque-roots.js:

Source/JavaScriptCore:

The Element object is a test object provided in the jsc shell for testing use only.
JavaScriptCore expects host constructors to either throw an error or return a
constructed object.  Element has a host constructor that did not obey this contract.
As a result, the following statement will fail a RELEASE_ASSERT:

    new (Element.bind())

This is now fixed.

* jsc.cpp:
(functionCreateElement):



Identifier: 178899@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204388 268f45cc-cd09-0410-ab3c-d52691b4dbfc