EWS logs file are rotated too quickly https://bugs.webkit.org/show_bug.cgi?id=160724
Reviewed by Daniel Bates.

Currently we are rotating the log file after 10 iterations of queue. If the queue
doesn't have any pending patches, these iterations result in very small amount of
logs (1 KB log file). Even if the queue process some patches, logs are few KBs.
This results in a lot of log files in a day. This patch ensures that we rotate the
log file when its file size is greater than or equal to 100 KB.

* EWSTools/start-queue-mac.sh: Added check for file size before rotating log file.


Identifier: 179023@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@204527 268f45cc-cd09-0410-ab3c-d52691b4dbfc