[css-grid] Do not recursively call layout during auto repeat computation https://bugs.webkit.org/show_bug.cgi?id=161234
Reviewed by Manuel Rego Casasnovas.

Source/WebCore:

The computation of auto repeat tracks was incorrectly recursively triggering
a layout in order to compute the available size. That was happening whenever
the width was indefinite. In such cases we should treat the width always as
indefinite without having to run any extra code. During the layout phase
we'll have the actual width available.

This fix also unveiled a few errors in one of the imported mozilla tests which were
incorrectly passing when they should not.

Test: fast/css-grid-layout/grid-indefinite-size-auto-repeat-crash.html

* rendering/RenderGrid.cpp:
(WebCore::RenderGrid::layoutBlock):
(WebCore::RenderGrid::computeIntrinsicLogicalWidths):
(WebCore::RenderGrid::computeAutoRepeatTracksCount):
(WebCore::RenderGrid::placeItemsOnGrid):
* rendering/RenderGrid.h:

LayoutTests:

Added a new test case and fixed a few tests inside the imported
grid-repeat-auto-fill-fit-001.html which were wrong due to a bug in Firefox. The intrinsic
size computation was not considering the existence of definite minimum sizes for some of the
grids so the outcome was narrower than expected.

* fast/css-grid-layout/grid-indefinite-size-auto-repeat-crash-expected.txt: Added.
* fast/css-grid-layout/grid-indefinite-size-auto-repeat-crash.html: Added.
* fast/css-grid-layout/mozilla/grid-repeat-auto-fill-fit-001-expected.html:

Identifier: 179488@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@205114 268f45cc-cd09-0410-ab3c-d52691b4dbfc