ASSERTION FAILED: !m_committedWidth in WebCore::LineWidth::fitBelowFloats https://bugs.webkit.org/show_bug.cgi?id=149462 <rdar://problem/27710841>
Reviewed by David Hyatt.

Source/WebCore:

In certain cases (multiple spans on the same line with negativ marings), the LineWidth::m_committedWidth > 0
check is not sufficient to decide if some content has already been committed to the current line.
This patch adds a flag to indicate if we ever committed to the current line.

Test: fast/text/assert-when-text-with-negative-margin-sibling-does-not-fit.html

* rendering/line/BreakingContext.h:
(WebCore::BreakingContext::handleText):
* rendering/line/LineWidth.h:
(WebCore::LineWidth::hasCommitted):

LayoutTests:

* fast/text/assert-when-text-with-negative-margin-sibling-does-not-fit-expected.txt: Added.
* fast/text/assert-when-text-with-negative-margin-sibling-does-not-fit.html: Added.


Identifier: 179706@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@205374 268f45cc-cd09-0410-ab3c-d52691b4dbfc