Fix API tests after r205580 https://bugs.webkit.org/show_bug.cgi?id=161668
No new tests, but this fixes API tests on bots.

* platform/URLParser.cpp:
(WebCore::domainToASCII):
This function seems to be appending characters to strings sometimes on some configurations.
This definitely needs further investigation, but this will fix the bots, and nothing else
will be affected because the URLParser is disabled by default.



Identifier: 179897@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@205582 268f45cc-cd09-0410-ab3c-d52691b4dbfc